智能找房

+++ core-model/src/main/java/com/toutiao/web/domain/query/ProjHouseInfoResponse.java
@@ -167,7 +167,7 @@ public class ProjHouseInfoResponse {
     /**
      * 是否删除(0-未删除/1-已删除)
      */
-    private Short isDel;
+    private int isDel;
 
 
     /**
